Moved api error to its crate and removed warnings
