+++ server/src/server_data/models/mod.rs
@@ -1,4 +1,3 @@
-pub mod api_error;
 pub mod command;
 pub mod company;
 pub mod history;
