+++ server/src/server_data/repos/connection.rs
@@ -1,4 +1,3 @@
-#[macro_use]
 use diesel::pg::PgConnection;
 use diesel::prelude::*;
 use diesel::r2d2::{ConnectionManager, Pool, PoolError, PooledConnection};
