+++ server/src/server_data/schema.rs
@@ -15,7 +15,6 @@ table! {
 
 table! {
     use diesel::sql_types::*;
-    use crate::models::command::*;
 
     company (id) {
         id -> Int4,
@@ -26,7 +25,6 @@ table! {
 
 table! {
     use diesel::sql_types::*;
-    use crate::models::command::*;
 
     history (id) {
         id -> Int4,
@@ -40,7 +38,6 @@ table! {
 
 table! {
     use diesel::sql_types::*;
-    use crate::models::command::*;
 
     news (id) {
         id -> Int4,
@@ -53,7 +50,6 @@ table! {
 
 table! {
     use diesel::sql_types::*;
-    use crate::models::command::*;
 
     stock (id) {
         id -> Int4,
@@ -67,7 +63,6 @@ table! {
 
 table! {
     use diesel::sql_types::*;
-    use crate::models::command::*;
 
     stonker (id) {
         id -> Int4,
